services/RecommendService/requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
langdetect
